Movie details

+++ MovieAPI/Models/MovieDetails.cs
@@ -6,5 +6,5 @@ public class MovieDetails
     public int MovieId { get; set; }
     public string Synopsis { get; set; } = string.Empty;
     public string Language { get; set; } = string.Empty; 
-    public int Budget { get; set; }
+    public float Budget { get; set; }
 }
